Revert "Make the overlay work again."
This reverts commit e5ef88e3ecff89cc72649d0548efbdb0ae79f4c0.

The overlay was made to be visible only in the static snapshot, and to be shown
only after the script has checked that it hasn't already been dismissed, to
prevent it appearing and disappearing briefly during load.

The reverted change made it such that the script only runs under ReSpec, which
makes the overlay always visible in the static spec and impossible to dismiss.