Link to html.spec.whatwg.org for needed HTML terms
http://validator.w3.org/checklink currently complains about all of these
links, because the anchors aren't actually in the linked page, instead
there's a script that forwards to a page of the multi-page spec.

To appease the link checker, one must either link to the single-page
spec or to a specific page of the multi-page spec.

To complicate matters, WHATWG and W3C are no longer in close sync, so
there's one reference that differs: #attr-class has become #classes.

Pick html.spec.whatwg.org to see broken references earlier.