Require the boxes in boxes, not output, to be within the title area box
This is almost equivalent as the "done positioning" step removes any
line boxes that do not completely fit inside the video, but as actually
implemented in at least Presto and Blink, it's the boxes in boxes which
are checked, not all of the boxes in output. Make this explicitly so.