Support all of HTML's character references
This hooks in to HTML's 'consume a character reference' for parsing,
and 'character references' for the syntax. All forms of charrefs
that HTML supports are supported, including missing semicolon.
(But missing semicolon is not valid, nor is a lone ampersand.)

Fixes https://www.w3.org/Bugs/Public/show_bug.cgi?id=23867.
Ref. #198.