Make VTTRegion lines be a long.
Because WebIDL doesn't include an unsigned int, the VTTRegion interface
specifically calls out that negative numbers should cause an
IndexSizeError, and browser implementations are in the unsigned int,
switch over to using a regular long value.
Using a long gives us the correct range and with the definition that
specifies throwing means that negative values are not allowed.

This was discussed in the [20 June, 2019 TTWG call](https://www.w3.org/2019/06/20-tt-minutes.html#x02)

Fixes #457.