[Editorial] Add "noexport" to :past and :future dfns (#510)
The `:past` and `:future` pseudo-classes are formally defined in Selectors 4,
which the WebVTT spec correctly references. It is fine to re-define these
terms locally. The problem is that previous commit flagged these definitions
as being selectors, which is correct but also made Bikeshed export them.

In other words, these definitions now conflict with those in Selectors 4.

This update flags these definitions with a `noexport` attribute to tell
Bikeshed to keep these definitions local.