Add support for ::cue-container
Fixes #516

In order to align the capabilities of WebVTT with CEA standards like 708,
add the concept of a "WebVTT block backgound box", which is defined as
a block-level wrapper around the WebVTT cue's contents.

To allow authors to specify styles that apply to this block background
object, add a new selector `::cue-container` that matches it. UAs which
provide FCC required styling information for "window background & opacity"
can do so via this new selector in UA-provided style sheets.